Create Main Screen code
The main screen code consists of arrays and screen navigation coding.  The main screen also has the main parts of the code which are the User Interface, the exit button, the button used to switch screens to go to the second screen and a few other things which are explained in the code.