Create Screen two
This code consists of all the coding which was used in the second screen. This code has functions, loops and arrays. It also has the coding for the second screens User interface coding and the code which will take you back to screen 1. After each section of the code there are comments which help you understand each part of the code.